+++ python/MaterialXTest/main.py
@@ -18,6 +18,8 @@ _testValues = (1,
                mx.Vector2(1.0, 2.0),
                mx.Vector3(1.0, 2.0, 3.0),
                mx.Vector4(1.0, 2.0, 3.0, 4.0),
+               mx.Matrix3x3(0.0),
+               mx.Matrix4x4(1.0),
                'value')
 
 _fileDir = os.path.dirname(os.path.abspath(__file__))
@@ -25,7 +27,8 @@ _libraryDir = os.path.join(_fileDir, '../../documents/Libraries/')
 _exampleDir = os.path.join(_fileDir, '../../documents/Examples/')
 _searchPath = _libraryDir + ';' + _exampleDir
 
-_libraryFilename = 'mx_stdlib_defs.mtlx'
+_libraryFilenames = ('mx_stdlib_defs.mtlx',
+                     'mx_stdlib_impl_osl.mtlx')
 _exampleFilenames = ('CustomNode.mtlx',
                      'Looks.mtlx',
                      'MaterialGraphs.mtlx',
@@ -77,27 +80,41 @@ class TestMaterialX(unittest.TestCase):
         self.assertTrue(doc.validate()[0])
 
         # Test scoped attributes.
+        nodeGraph.setFilePrefix('folder/')
         nodeGraph.setColorSpace('lin_rec709')
-        nodeGraph.setFilePrefix('prefix')
+        self.assertTrue(image.getParameter('file').getResolvedValueString() == 'folder/image1.tif')
         self.assertTrue(constant.getActiveColorSpace() == 'lin_rec709')
-        self.assertTrue(image.getActiveFilePrefix() == 'prefix')
 
         # Create a simple shader interface.
-        shader = doc.addNodeDef('shader1', 'surfaceshader', 'simpleSrf')
-        diffColor = shader.addInput('diffColor', 'color3')
-        specColor = shader.addInput('specColor', 'color3')
-        roughness = shader.addParameter('roughness', 'float')
+        shaderDef = doc.addNodeDef('shader1', 'surfaceshader', 'simpleSrf')
+        diffColor = shaderDef.setInputValue('diffColor', mx.Color3(1.0))
+        specColor = shaderDef.setInputValue('specColor', mx.Color3(0.0))
+        roughness = shaderDef.setParameterValue('roughness', 0.25)
+        self.assertTrue(roughness.getValue() == 0.25)
 
         # Create a material that instantiates the shader.
         material = doc.addMaterial()
-        shaderRef = material.addShaderRef('', 'simpleSrf')
-        self.assertTrue(material.getReferencedShaderDefs())
+        shaderRef = material.addShaderRef('shaderRef1', 'simpleSrf')
+        self.assertTrue(roughness.getBoundValue(material) == 0.25)
 
-        # Bind the diffuse color input to the constant color output.
+        # Bind a shader input to a value.
+        bindInput = shaderRef.addBindInput('specColor')
+        bindInput.setValue(mx.Color3(0.5))
+        self.assertTrue(specColor.getBoundValue(material) == mx.Color3(0.5))
+        self.assertTrue(specColor.getDefaultValue() == mx.Color3(0.0))
+
+        # Bind a shader parameter to a value.
+        bindParam = shaderRef.addBindParam('roughness')
+        bindParam.setValue(0.5)
+        self.assertTrue(roughness.getBoundValue(material) == 0.5)
+        self.assertTrue(roughness.getDefaultValue() == 0.25)
+
+        # Bind a shader input to a graph output.
         bindInput = shaderRef.addBindInput('diffColor')
-        bindInput.setConnectedOutput(output1)
-        self.assertTrue(output1 in shaderRef.getReferencedOutputs())
-        self.assertTrue(diffColor.getUpstreamElement(material) == output1)
+        bindInput.setConnectedOutput(output2)
+        self.assertTrue(diffColor.getUpstreamElement(material) == output2)
+        self.assertTrue(diffColor.getBoundValue(material) is None)
+        self.assertTrue(diffColor.getDefaultValue() == mx.Color3(1.0))
 
         # Create a collection.
         collection = doc.addCollection()
@@ -245,9 +262,12 @@ class TestMaterialX(unittest.TestCase):
 
     def test_ReadXml(self):
         # Load the standard library.
-        lib = mx.createDocument()
-        mx.readFromXmlFile(lib, _libraryFilename, _searchPath)
-        self.assertTrue(lib.validate()[0])
+        libs = []
+        for filename in _libraryFilenames:
+            lib = mx.createDocument()
+            mx.readFromXmlFile(lib, filename, _searchPath)
+            self.assertTrue(lib.validate()[0])
+            libs.append(lib)
 
         for filename in _exampleFilenames:
             # Read the example document.
@@ -270,15 +290,19 @@ class TestMaterialX(unittest.TestCase):
 
             # Traverse upstream from each shader input.
             for material in doc.getMaterials():
-                self.assertTrue(material.getReferencedShaderDefs())
+                self.assertTrue(material.getPrimaryShaderNodeDef())
                 edgeCount = 0
-                for shader in material.getReferencedShaderDefs():
-                    for input in shader.getInputs():
-                        for edge in input.traverseGraph(material):
-                            edgeCount += 1
-                    for param in shader.getParameters():
-                        for edge in param.traverseGraph(material):
-                            edgeCount += 1
+                for param in material.getPrimaryShaderParameters():
+                    boundValue = param.getBoundValue(material)
+                    self.assertTrue(boundValue is not None)
+                    for edge in param.traverseGraph(material):
+                        edgeCount += 1
+                for input in material.getPrimaryShaderInputs():
+                    boundValue = input.getBoundValue(material)
+                    upstreamElement = input.getUpstreamElement(material)
+                    self.assertTrue(boundValue is not None or upstreamElement is not None)
+                    for edge in input.traverseGraph(material):
+                        edgeCount += 1
                 self.assertTrue(edgeCount > 0)
 
             # Serialize to XML.
@@ -291,13 +315,15 @@ class TestMaterialX(unittest.TestCase):
 
             # Combine document with the standard library.
             doc2 = doc.copy()
-            doc2.importLibrary(lib)
+            for lib in libs:
+                doc2.importLibrary(lib)
             self.assertTrue(doc2.validate()[0])
 
-            # Verify that all referenced nodes are declared.
+            # Verify that all referenced nodes are declared and implemented.
             for elem in doc2.traverseTree():
                 if elem.isA(mx.Node):
-                    self.assertTrue(elem.getReferencedNodeDef())
+                    self.assertTrue(elem.getNodeDef())
+                    self.assertTrue(elem.getImplementation())
 
 
 #--------------------------------------------------------------------------------