+++ source/MaterialXCore/Interface.cpp
@@ -61,22 +61,25 @@ bool PortElement::validate(string* message) const
 // Parameter methods
 //
 
-Edge Parameter::getUpstreamEdge(ConstMaterialPtr material, size_t index)
+Edge Parameter::getUpstreamEdge(ConstMaterialPtr material, size_t index) const
 {
     if (material && index < getUpstreamEdgeCount())
     {
-        if (getParent()->isA<NodeDef>())
+        ConstNodeDefPtr nodeDef = getParent()->isA<Implementation>() ?
+                                  getParent()->asA<Implementation>()->getNodeDef() :
+                                  getParent()->asA<NodeDef>();
+        if (nodeDef)
         {
             // Apply BindParam elements to the Parameter.
             for (ShaderRefPtr shaderRef : material->getShaderRefs())
             {
-                if (shaderRef->getReferencedShaderDef() == getParent())
+                if (shaderRef->getNodeDef() == nodeDef)
                 {
                     for (BindParamPtr bindParam : shaderRef->getBindParams())
                     {
                         if (bindParam->getName() == getName() && bindParam->hasValue())
                         {
-                            return Edge(getSelf(), nullptr, bindParam);
+                            return Edge(getSelfNonConst(), nullptr, bindParam);
                         }
                     }
                 }
@@ -87,7 +90,7 @@ Edge Parameter::getUpstreamEdge(ConstMaterialPtr material, size_t index)
         OverridePtr override = material->getOverride(getPublicName());
         if (override)
         {
-            return Edge(getSelf(), nullptr, override);
+            return Edge(getSelfNonConst(), nullptr, override);
         }
     }
 
@@ -98,18 +101,21 @@ Edge Parameter::getUpstreamEdge(ConstMaterialPtr material, size_t index)
 // Input methods
 //
 
-Edge Input::getUpstreamEdge(ConstMaterialPtr material, size_t index)
+Edge Input::getUpstreamEdge(ConstMaterialPtr material, size_t index) const
 {
     if (material && index < getUpstreamEdgeCount())
     {
-        if (getParent()->isA<NodeDef>())
+        ConstNodeDefPtr nodeDef = getParent()->isA<Implementation>() ?
+                                  getParent()->asA<Implementation>()->getNodeDef() :
+                                  getParent()->asA<NodeDef>();
+        if (nodeDef)
         {
             if (material)
             {
                 // Apply BindInput elements to the Input.
                 for (ShaderRefPtr shaderRef : material->getShaderRefs())
                 {
-                    if (shaderRef->getReferencedShaderDef() == getParent())
+                    if (shaderRef->getNodeDef() == nodeDef)
                     {
                         for (BindInputPtr bindInput : shaderRef->getBindInputs())
                         {
@@ -120,11 +126,11 @@ Edge Input::getUpstreamEdge(ConstMaterialPtr material, size_t index)
                             OutputPtr output = bindInput->getConnectedOutput();
                             if (output)
                             {
-                                return Edge(getSelf(), bindInput, output);
+                                return Edge(getSelfNonConst(), bindInput, output);
                             }
                             if (bindInput->hasValue())
                             {
-                                return Edge(getSelf(), nullptr, bindInput);
+                                return Edge(getSelfNonConst(), nullptr, bindInput);
                             }
                         }
                     }
@@ -136,7 +142,7 @@ Edge Input::getUpstreamEdge(ConstMaterialPtr material, size_t index)
         OverridePtr override = material->getOverride(getPublicName());
         if (override)
         {
-            return Edge(getSelf(), nullptr, override);
+            return Edge(getSelfNonConst(), nullptr, override);
         }
     }
 
@@ -147,11 +153,11 @@ Edge Input::getUpstreamEdge(ConstMaterialPtr material, size_t index)
 // Output methods
 //
 
-Edge Output::getUpstreamEdge(ConstMaterialPtr material, size_t index)
+Edge Output::getUpstreamEdge(ConstMaterialPtr material, size_t index) const
 {
     if (index < getUpstreamEdgeCount())
     {
-        return Edge(getSelf(), nullptr, getConnectedNode());
+        return Edge(getSelfNonConst(), nullptr, getConnectedNode());
     }
 
     return NULL_EDGE;
@@ -181,16 +187,16 @@ bool Output::validate(string* message) const
 // InterfaceElement methods
 //
 
-const string& InterfaceElement::getParameterValueString(const string& name) const
-{
-    ParameterPtr param = getChildOfType<Parameter>(name);
-    return param ? param->getValueString() : EMPTY_STRING;
-}
-
 ValuePtr InterfaceElement::getParameterValue(const string& name) const
 {
     ParameterPtr param = getChildOfType<Parameter>(name);
     return param ? param->getValue() : ValuePtr();
+}
+
+ValuePtr InterfaceElement::getInputValue(const string& name) const
+{
+    InputPtr input = getChildOfType<Input>(name);
+    return input ? input->getValue() : ValuePtr();
 }
 
 void InterfaceElement::registerChildElement(ElementPtr child)
@@ -217,6 +223,31 @@ void InterfaceElement::unregisterChildElement(ElementPtr child)
     {
         _inputCount--;
     }
+}
+
+bool InterfaceElement::isTypeCompatible(InterfaceElementPtr rhs) const
+{
+    if (getType() != rhs->getType())
+    {
+        return false;
+    }
+    for (ParameterPtr param : getParameters())
+    {
+        ParameterPtr matchingParam = rhs->getParameter(param->getName());
+        if (matchingParam && matchingParam->getType() != param->getType())
+        {
+            return false;
+        }
+    }
+    for (InputPtr input : getInputs())
+    {
+        InputPtr matchingInput = rhs->getInput(input->getName());
+        if (matchingInput && matchingInput->getType() != input->getType())
+        {
+            return false;
+        }
+    }
+    return true;
 }
 
 } // namespace MaterialX