+++ source/MaterialXCore/Material.cpp
@@ -25,23 +25,28 @@ ShaderRefPtr Material::addShaderRef(const string& name, const string& node)
     ShaderRefPtr shaderRef = addChild<ShaderRef>(name);
     if (!node.empty())
     {
-        shaderRef->setNode(node);
+        shaderRef->setNodeString(node);
     }
     return shaderRef;
 }
 
-vector<NodeDefPtr> Material::getReferencedShaderDefs() const
+vector<NodeDefPtr> Material::getShaderNodeDefs(const string& target, const string& type) const
 {
-    vector<NodeDefPtr> shaderDefs;
+    vector<NodeDefPtr> nodeDefs;
     for (ShaderRefPtr shaderRef : getShaderRefs())
     {
-        NodeDefPtr shaderDef = shaderRef->getReferencedShaderDef();
-        if (shaderDef)
+        NodeDefPtr nodeDef = shaderRef->getNodeDef();
+        if (!nodeDef || !targetStringsMatch(nodeDef->getTarget(), target))
         {
-            shaderDefs.push_back(shaderDef);
+            continue;
         }
+        if (!type.empty() && type != nodeDef->getType())
+        {
+            continue;
+        }
+        nodeDefs.push_back(nodeDef);
     }
-    return shaderDefs;
+    return nodeDefs;
 }
 
 vector<MaterialAssignPtr> Material::getReferencingMaterialAssigns() const
@@ -125,15 +130,15 @@ OutputPtr BindInput::getConnectedOutput() const
 // ShaderRef methods
 //
 
-NodeDefPtr ShaderRef::getReferencedShaderDef()
+NodeDefPtr ShaderRef::getNodeDef()
 {
-    if (hasNodeDef())
+    if (hasNodeDefString())
     {
-        return getDocument()->getNodeDef(getNodeDef());
+        return getDocument()->getNodeDef(getNodeDefString());
     }
-    if (hasNode())
+    if (hasNodeString())
     {
-        vector<NodeDefPtr> nodeDefs = getDocument()->getMatchingNodeDefs(getNode());
+        vector<NodeDefPtr> nodeDefs = getDocument()->getMatchingNodeDefs(getNodeString());
         return nodeDefs.empty() ? NodeDefPtr() : nodeDefs[0];
     }
     return NodeDefPtr();