+++ source/MaterialXCore/Value.h
@@ -18,6 +18,8 @@ namespace MaterialX
 
 /// A shared pointer to a Value
 using ValuePtr = shared_ptr<class Value>;
+/// A shared pointer to a const Value
+using ConstValuePtr = shared_ptr<const class Value>;
 
 template <class T> class TypedValue;